+++ src/main/java/net/brnbrd/delightful/common/block/CantaloupeBlock.java
@@ -2,6 +2,7 @@ package net.brnbrd.delightful.common.block;
 
 import net.brnbrd.delightful.Util;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
@@ -10,8 +11,8 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
-import org.jetbrains.annotations.NotNull;
 import vectorwing.farmersdelight.common.tag.ForgeTags;
+import org.jetbrains.annotations.NotNull;
 
 public class CantaloupeBlock extends MiniBlock {
 	public CantaloupeBlock(Properties pProperties) {
@@ -20,16 +21,23 @@ public class CantaloupeBlock extends MiniBlock {
 
 	@SuppressWarnings("deprecation")
 	@Override
-	public @NotNull InteractionResult use(@NotNull BlockState pState, @NotNull Level world, @NotNull BlockPos pPos, Player pPlayer, @NotNull InteractionHand pHand, @NotNull BlockHitResult pHit) {
-		if (pPlayer.getItemInHand(pHand).is(ForgeTags.TOOLS_KNIVES)) {
-			if (!world.isClientSide()) {
+	public @NotNull InteractionResult use(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, Player player, @NotNull InteractionHand hand, @NotNull BlockHitResult hit) {
+		if (player.getItemInHand(hand).is(ForgeTags.TOOLS_KNIVES)) {
+			if (!level.isClientSide()) {
 				SlicedMiniMelonBlock sliced = (SlicedMiniMelonBlock) DelightfulBlocks.SLICED_CANTALOUPE.get();
-				world.setBlock(pPos, sliced.defaultBlockState(), 2);
-				Util.dropOrGive(sliced.getSliceItem(), world, pPos, pPlayer);
-				world.playSound(null, pPos, SoundEvents.WOOD_HIT, SoundSource.PLAYERS, 0.8F, 0.8F);
-				pPlayer.getItemInHand(pHand).hurtAndBreak(1, pPlayer, onBroken -> onBroken.broadcastBreakEvent(pHand));
+				level.setBlock(pos, sliced.defaultBlockState(), 2);
+				Direction direction = hit.getDirection();
+				Util.dropOrGive(
+					sliced.getSliceItem(),
+					level,
+					pos,
+					direction.getAxis() == Direction.Axis.Y ? player.getDirection().getOpposite() : direction,
+					player
+				);
+				level.playSound(null, pos, SoundEvents.PUMPKIN_CARVE, SoundSource.BLOCKS, 1F, 1F);
+				player.getItemInHand(hand).hurtAndBreak(1, player, onBroken -> onBroken.broadcastBreakEvent(hand));
 			}
-			return InteractionResult.sidedSuccess(world.isClientSide());
+			return InteractionResult.sidedSuccess(level.isClientSide());
 		}
 		return InteractionResult.FAIL;
 	}