+++ src/main/java/net/brnbrd/delightful/common/block/SlicedMelonBlock.java
@@ -4,6 +4,7 @@ import com.mojang.datafixers.util.Pair;
 import net.brnbrd.delightful.Util;
 import net.brnbrd.delightful.common.DelightfulConfig;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
@@ -18,7 +19,6 @@ import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.MelonBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.IntegerProperty;
@@ -32,7 +32,7 @@ import org.jetbrains.annotations.NotNull;
 import vectorwing.farmersdelight.common.tag.ForgeTags;
 import java.util.function.Supplier;
 
-public class SlicedMelonBlock extends MelonBlock implements ISliceable {
+public class SlicedMelonBlock extends Block implements ISliceable {
 	public static final Integer MAX_BITES = 8;
 	public static final IntegerProperty BITES = IntegerProperty.create("bites", 1, MAX_BITES);
 	private final Supplier<Item> sliceItem;
@@ -83,7 +83,7 @@ public class SlicedMelonBlock extends MelonBlock implements ISliceable {
 
 	@Override
 	public ItemStack getSliceItem() {
-		return new ItemStack(this.sliceItem.get());
+		return Util.gs(this.sliceItem, 1);
 	}
 
 	public ItemStack getJuiceItem() {
@@ -100,7 +100,7 @@ public class SlicedMelonBlock extends MelonBlock implements ISliceable {
 	public @NotNull InteractionResult use(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, Player player, @NotNull InteractionHand hand, @NotNull BlockHitResult hit) {
 		ItemStack heldStack = player.getItemInHand(hand);
 		if (heldStack.is(ForgeTags.TOOLS_KNIVES)) {
-			return this.cutSlice(level, pos, state, player, hand);
+			return this.cutSlice(level, pos, state, player, hand, hit);
 		} else if (heldStack.is(Items.GLASS_BOTTLE) && DelightfulConfig.MELON_JUICING.get()) {
 			return this.bottleJuice(level, pos, state, player, hand);
 		}
@@ -129,12 +129,12 @@ public class SlicedMelonBlock extends MelonBlock implements ISliceable {
 			} else if (bites < this.getMaxBites()) {
 				level.setBlock(pos, state.setValue(BITES, bites + 1), 3);
 			}
-			level.playSound(null, pos, SoundEvents.GENERIC_EAT, SoundSource.PLAYERS, 0.8F, 0.8F);
+			level.playSound(null, playerIn.blockPosition().above(), SoundEvents.GENERIC_EAT, SoundSource.PLAYERS, 1F, 1F);
 		}
 		return InteractionResult.sidedSuccess(level.isClientSide());
 	}
 
-	protected InteractionResult cutSlice(Level level, BlockPos pos, BlockState state, Player player, InteractionHand hand) {
+	protected InteractionResult cutSlice(Level level, BlockPos pos, BlockState state, Player player, InteractionHand hand, BlockHitResult hit) {
 		if (!level.isClientSide()) {
 			int bites = state.getValue(BITES);
 			if (bites == this.getMaxBites()) {
@@ -142,8 +142,15 @@ public class SlicedMelonBlock extends MelonBlock implements ISliceable {
 			} else if (bites < this.getMaxBites()) {
 				level.setBlock(pos, state.setValue(BITES, bites + 1), 3);
 			}
-			Util.dropOrGive(this.getSliceItem(), level, pos, player);
-			level.playSound(null, pos, SoundEvents.WOOD_HIT, SoundSource.PLAYERS, 0.8F, 0.8F);
+			Direction direction = hit.getDirection();
+			Util.dropOrGive(
+				this.getSliceItem(),
+				level,
+				pos,
+				direction.getAxis() == Direction.Axis.Y ? player.getDirection().getOpposite() : direction,
+				player
+			);
+			level.playSound(null, pos, SoundEvents.PUMPKIN_CARVE, SoundSource.BLOCKS, 1F, 1F);
 			player.getItemInHand(hand).hurtAndBreak(1, player, onBroken -> onBroken.broadcastBreakEvent(hand));
 		}
 		return InteractionResult.sidedSuccess(level.isClientSide());
@@ -162,7 +169,7 @@ public class SlicedMelonBlock extends MelonBlock implements ISliceable {
 			}
 			player.getItemInHand(hand).shrink(1);
 			ItemHandlerHelper.giveItemToPlayer(player, this.getJuiceItem(), 0);
-			level.playSound(null, pos, SoundEvents.BOTTLE_FILL, SoundSource.PLAYERS, 0.8F, 0.8F);
+			level.playSound(null, pos, SoundEvents.BOTTLE_FILL, SoundSource.PLAYERS, 1F, 1F);
 		}
 		return InteractionResult.sidedSuccess(level.isClientSide());
 	}