+++ src/main/java/net/brnbrd/delightful/common/events/ForgeEvents.java
@@ -2,30 +2,27 @@ package net.brnbrd.delightful.common.events;
 
 import net.brnbrd.delightful.Util;
 import net.brnbrd.delightful.common.block.DelightfulBlocks;
-import net.brnbrd.delightful.common.block.SlicedGourdBlock;
-import net.brnbrd.delightful.common.block.SlicedMelonBlock;
+import net.brnbrd.delightful.common.block.ISliceable;
 import net.brnbrd.delightful.common.item.DelightfulItems;
 import net.brnbrd.delightful.compat.BrewinChewinCompat;
 import net.brnbrd.delightful.compat.CasualnessDelightCompat;
 import net.brnbrd.delightful.compat.Modid;
-import net.brnbrd.delightful.compat.Mods;
 import net.brnbrd.delightful.data.tags.DelightfulItemTags;
 import net.minecraft.core.BlockPos;
-import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.npc.VillagerTrades;
-import net.minecraft.world.item.Item;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.trading.MerchantOffer;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.properties.BlockStateProperties;
-import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraftforge.event.entity.living.LivingEntityUseItemEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.event.village.WandererTradesEvent;
@@ -38,7 +35,6 @@ import java.util.List;
 import java.util.Objects;
 
 public class ForgeEvents {
-
 	@SubscribeEvent
 	void onEatEffectProvider(LivingEntityUseItemEvent.Finish e) {
 		if (e.getResult() != Event.Result.DENY) {
@@ -65,70 +61,55 @@ public class ForgeEvents {
 	void onWanderingTrader(WandererTradesEvent e) {
 		List<VillagerTrades.ItemListing> trades = e.getGenericTrades();
 		if (Util.enabled(DelightfulItems.SALMONBERRIES) && Util.enabled(DelightfulItems.SALMONBERRY_PIPS)) {
-			trades.add((ent, r) -> new MerchantOffer(new ItemStack(Items.EMERALD, 1), Util.gs(DelightfulItems.SALMONBERRY_PIPS), 5, 1, 1));
+			trades.add((ent, r) -> new MerchantOffer(new ItemStack(Items.EMERALD, 2), Util.gs(DelightfulItems.SALMONBERRY_PIPS), 5, 1, 1));
 		}
 		if (Util.enabled(DelightfulItems.CANTALOUPE) && Util.enabled(DelightfulItems.CANTALOUPE_SLICE)) {
-			trades.add((ent, r) -> new MerchantOffer(new ItemStack(Items.EMERALD, 2), Util.gs(DelightfulItems.CANTALOUPE_SLICE, 8), 5, 1, 1));
+			trades.add((ent, r) -> new MerchantOffer(new ItemStack(Items.EMERALD, 3), Util.gs(DelightfulItems.CANTALOUPE_SLICE, 8), 5, 1, 1));
+		}
+		if (
+			Util.enabled(DelightfulItems.CANTALOUPE) &&
+			Util.enabled(DelightfulItems.STUFFED_CANTALOUPE_BLOCK) &&
+			Util.enabled(DelightfulItems.STUFFED_CANTALOUPE)
+		) {
+			trades.add((ent, r) -> new MerchantOffer(new ItemStack(Items.EMERALD, 5), Util.gs(DelightfulItems.STUFFED_CANTALOUPE_BLOCK, 1), 3, 1, 1));
 		}
 	}
 
-	// Right slick slicing an Item from a Block
+	// Crouch right-click slicing
 	@SubscribeEvent(priority = EventPriority.HIGHEST)
 	void onInteract(PlayerInteractEvent.RightClickBlock e) {
-		Level world = e.getLevel();
+		Level level = e.getLevel();
 		BlockPos pos = e.getPos();
-		if (e.getItemStack().is(ForgeTags.TOOLS_KNIVES)) {
-			BlockState current = world.getBlockState(pos);
-			boolean client = world.isClientSide();
-			if (current.getBlock() == Blocks.MELON) {
-				SlicedMelonBlock sliced = (SlicedMelonBlock) DelightfulBlocks.SLICED_MELON.get();
-				slice(sliced.defaultBlockState(), sliced.getSliceItem(), world, pos, SoundEvents.BAMBOO_BREAK, e, client);
-			} else if (current.getBlock() == Blocks.PUMPKIN && !e.getEntity().isCrouching()) {
-				SlicedGourdBlock sliced = (SlicedGourdBlock) DelightfulBlocks.SLICED_PUMPKIN.get();
-				slice(sliced.defaultBlockState(), sliced.getSliceItem(), world, pos, SoundEvents.BAMBOO_BREAK, e, client);
+		if (e.getItemStack().is(ForgeTags.TOOLS_KNIVES) && !e.getEntity().isCrouching()) {
+			Block current = level.getBlockState(pos).getBlock();
+			Block newBlock;
+			if (current == Blocks.MELON) {
+				newBlock = DelightfulBlocks.SLICED_MELON.get();
+			} else if (current == Blocks.PUMPKIN) {
+				newBlock = DelightfulBlocks.SLICED_PUMPKIN.get();
 			} else if (
-				Modid.UGD.loaded() &&
-				e.getEntity().isCrouching() && // Must be crouching to avoid "carving"
-				Objects.equals(ForgeRegistries.BLOCKS.getKey(current.getBlock()), Modid.UG.rl("gloomgourd"))
+				Objects.equals(ForgeRegistries.BLOCKS.getKey(current), Modid.UG.rl("carved_gloomgourd"))
 			) {
-				SlicedGourdBlock sliced = (SlicedGourdBlock) DelightfulBlocks.SLICED_GLOOMGOURD.get();
-				slice(sliced.defaultBlockState(), sliced.getSliceItem(), world, pos, SoundEvents.BAMBOO_BREAK, e, client);
-			} else if (
-				Mods.loaded(Modid.FU, Modid.FUD) &&
-				Util.name(current.getBlock()).equals("truffle_cake") &&
-				Modid.FUD.itemExists("truffle_cake_slice")
-			) {
-				Item sliceItem = Modid.FUD.item("truffle_cake_slice");
-				if (sliceItem != null) {
-					int currentBites = current.getValue(BlockStateProperties.BITES);
-					ItemStack slice = sliceItem.getDefaultInstance();
-					if (currentBites >= 3) {
-						world.removeBlock(pos, false);
-						world.gameEvent(e.getEntity(), GameEvent.BLOCK_DESTROY, pos);
-						Util.dropOrGive(slice, world, pos, e.getEntity());
-						world.playSound(null, pos, SoundEvents.WOOL_PLACE, SoundSource.PLAYERS, 0.8F, 0.8F);
-						e.getEntity().getItemInHand(e.getHand()).hurtAndBreak(1, e.getEntity(), onBroken -> {
-						});
-						e.setCancellationResult(InteractionResult.sidedSuccess(client));
-						e.setCanceled(true);
-						return;
-					}
-					slice(current.setValue(BlockStateProperties.BITES, currentBites + 1), slice, world, pos, SoundEvents.WOOL_PLACE, e, client);
+				newBlock = DelightfulBlocks.SLICED_GLOOMGOURD.get();
+			} else {
+				return;
+			}
+			if (newBlock instanceof ISliceable slice) {
+				boolean client = level.isClientSide();
+				e.setCancellationResult(InteractionResult.sidedSuccess(client));
+				e.setCanceled(true);
+				if (!client) {
+					slice(newBlock.defaultBlockState(), slice.getSliceItem(), level, pos, e.getEntity(), e.getHand());
 				}
 			}
 		}
 	}
 
 	// Replaces Block in world, drops Item, cancels interaction event
-	void slice(BlockState block, ItemStack slice, Level world, BlockPos pos, SoundEvent sound, PlayerInteractEvent.RightClickBlock e, boolean client) {
-		if (!client) {
-			world.setBlock(pos, block, 2);
-			Util.dropOrGive(slice, world, pos, e.getEntity());
-			world.playSound(null, pos, sound, SoundSource.PLAYERS, 0.8F, 0.8F);
-			e.getEntity().getItemInHand(e.getHand()).hurtAndBreak(1, e.getEntity(), onBroken -> {
-			});
-		}
-		e.setCancellationResult(InteractionResult.sidedSuccess(client));
-		e.setCanceled(true);
+	void slice(BlockState block, ItemStack slice, Level world, BlockPos pos, Player player, InteractionHand hand) {
+		world.setBlock(pos, block, 2);
+		Util.dropOrGive(slice, world, pos, player);
+		world.playSound(null, pos, SoundEvents.PUMPKIN_CARVE, SoundSource.BLOCKS, 1F, 1F);
+		player.getItemInHand(hand).hurtAndBreak(1, player, onBroken -> {});
 	}
 }