+++ src/main/java/net/brnbrd/delightful/compat/undergarden/UndergardenCompat.java
@@ -1,6 +1,7 @@
-package net.brnbrd.delightful.compat;
+package net.brnbrd.delightful.compat.undergarden;
 
 import net.brnbrd.delightful.Util;
+import net.brnbrd.delightful.compat.Modid;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
@@ -8,7 +9,6 @@ import net.minecraft.world.food.FoodProperties;
 import java.util.function.Supplier;
 
 public class UndergardenCompat {
-
 	public static MobEffect getVirulentResistance() {
 		return Util.effect(Modid.UG, "virulent_resistance", MobEffects.FIRE_RESISTANCE);
 	}